// src/components/ui/profile-menu/type.ts
export interface User {
  lastName: string;
  firstName: string;
  middleName: string;
  region: string;
  city: string;
  street: string;
  house: string;
  building: string;
  apartment: string;
  index: string;
  email: string;
  phone: string;
  repeatPassword: string;
  oldPassword: string;
  newPassword: string;
}